+++ packages/form/lib/src/validation/rules/max_value.dart
@@ -1,10 +1,10 @@
-import 'package:bond_form/src/validation/validation_rule.dart';
 import 'package:bond_form/src/form_fields/form_field_state.dart';
+import 'package:bond_form/src/validation/validation_rule.dart';
 
 /// A validation rule that checks if a numeric value does not exceed a maximum limit.
 ///
 /// This rule validates that the numeric value does not exceed the specified maximum limit.
-class MaxValue extends ValidationRule<num> {
+class MaxValue extends ValidationRule<String> {
   /// The maximum allowed numeric value.
   final int max;
 
@@ -20,7 +20,13 @@ class MaxValue extends ValidationRule<num> {
       l10n.maxValueValidationMessage(fieldName, max);
 
   @override
-  bool validate(num value, Map<String, FormFieldState> fields) {
-    return value <= max;
+  bool validate(String value, Map<String, FormFieldState> fields) {
+    // Attempt to parse the value as an integer.
+    final parsedValue = num.tryParse(value);
+    // If parsing fails, the value is invalid.
+    if (parsedValue == null) {
+      return false;
+    }
+    return parsedValue <= max;
   }
 }