+++ packages/form/lib/src/validation/rules/min_value.dart
@@ -1,10 +1,10 @@
-import 'package:bond_form/src/validation/validation_rule.dart';
 import 'package:bond_form/src/form_fields/form_field_state.dart';
+import 'package:bond_form/src/validation/validation_rule.dart';
 
 /// A validation rule that checks if a numeric value meets a minimum requirement.
 ///
 /// This rule validates that the numeric value meets the specified minimum requirement.
-class MinValue extends ValidationRule<num> {
+class MinValue extends ValidationRule<String> {
   /// The minimum required numeric value.
   final int min;
 
@@ -20,7 +20,13 @@ class MinValue extends ValidationRule<num> {
       l10n.minValueValidationMessage(fieldName, min);
 
   @override
-  bool validate(num value, Map<String, FormFieldState> fields) {
-    return value >= min;
+  bool validate(String value, Map<String, FormFieldState> fields) {
+    // Attempt to parse the value as an integer.
+    final parsedValue = num.tryParse(value);
+    // If parsing fails, the value is invalid.
+    if (parsedValue == null) {
+      return false;
+    }
+    return parsedValue >= min;
   }
 }