+++ packages/form/packages/bond_form_getx/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+## 0.0.7
+* Updated bond_form dependency to ^0.0.14
 
 ## 0.0.6